requirements.txt
python-telegram-bot
ccxt
pandas
numpy
pyyaml
apscheduler
python-dotenv
pandas-ta